requirements.txt
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.95.0
requests==2.31.0
boto3==1.28.40
pytz==2023.3